+++ src/scitaste/data/__init__.py
@@ -1,16 +1,20 @@
 """Separate factual knowledge and scientific-taste data stores."""
 
-from scitaste.data.ingestion import ingest_corpus, normalize_corpus
+from scitaste.data.curation import CurationFormat, curate_snapshot
+from scitaste.data.ingestion import audit_corpus_manifest, ingest_corpus, normalize_corpus
 from scitaste.data.models import KnowledgeDocument, ProvenanceRecord, TasteCase
 from scitaste.data.store import KnowledgeLibrary, TasteLibrary, build_libraries
 
 __all__ = [
+    "CurationFormat",
     "KnowledgeDocument",
     "KnowledgeLibrary",
     "ProvenanceRecord",
     "TasteCase",
     "TasteLibrary",
+    "audit_corpus_manifest",
     "build_libraries",
+    "curate_snapshot",
     "ingest_corpus",
     "normalize_corpus",
 ]